(chore:NewTickets) rename disabledPreviewButton->previewButtonDisabled
